+++ packages/nimble-components/src/select/testing/select.pageobject.ts
@@ -93,15 +93,20 @@ export class SelectPageObject {
     }
 
     /**
-     * Click the option with the text provided by the 'value' parameter.
+     * Click the option with the text provided by the 'displayText' parameter.
      * @param value The text of the option to be selected
      */
-    public clickOptionWithDisplayText(value: string): void {
+    public async clickOptionWithDisplayText(
+        displayText: string
+    ): Promise<void> {
+        if (!this.selectElement.open) {
+            await this.clickSelect();
+        }
         const optionIndex = this.selectElement.options.findIndex(
-            o => o.text === value
+            o => o.text === displayText
         );
         if (optionIndex === -1) {
-            throw new Error(`No option with "text" of ${value}`);
+            throw new Error(`No option with "text" of ${displayText}`);
         }
 
         this.clickOption(optionIndex);